seminars/seminar4-event-driven/queues/load_test/requirements.txt
aiohttp==3.9.1
pillow==10.1.0
